+++ src/cmds/shell.js
@@ -58,10 +58,8 @@ class Shell extends CliTrackerController {
       };
 
       // Support extra envs, ports and mount volumes
-      cmd_options.envs   = this._parse_option(options.env  , /.+=.+/, '=', 'invalid_env');
-      if (!!cmd_options.interactive && !cmd_options.envs.TERM) {
-        cmd_options.envs.TERM = process.env.TERM;
-      }
+      cmd_options.envs       = this._parse_option(options.env  , /.+=.+/, '=', 'invalid_env');
+      cmd_options.shell_term = process.env.TERM;
       cmd_options.mounts = this._parse_option(options.mount, /.+:.+:?.*/, ':', 'invalid_mount', 1, (opts) => {
         return { type: (opts[2] ? opts[1] : 'path'), value: (opts[2] ? opts[2] : opts[0]) };
       });